Renaming global_player.lue to old
Appearantely EQEMU doesn't like to load the same functionality as a .pl
and .lua. So when we added a global_player.lua file, it stopped reading
the global_player.pl file.

The global_player.pl file is very important as it sets up all our zone
restrictions and what not. Disabling the .lua version until further
notice.